requirements.txt
Django==5.0.1
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
scikit-learn==1.4.0
torch==2.1.2
albumentations==1.3.1
gunicorn==21.2.0
djangorestframework==3.14.0
